Reorganizing aggregation logic
