+++ src/resolvers-cassandra/Edges/queries.js
@@ -4,27 +4,9 @@ const Promise = require('promise');
 const Long = require('cassandra-driver').types.Long;
 const cassandraConnector = require('../../clients/cassandra/CassandraConnector');
 const featureServiceClient = require('../../clients/locations/FeatureServiceClient');
-const { tilesForBbox, parseFromToDate, withRunTime, toPipelineKey, computeWeightedAvg, toConjunctionTopics, fromTopicListToConjunctionTopics } = require('../shared');
+const { tilesForBbox, parseFromToDate, withRunTime, toPipelineKey, aggregateBy, toConjunctionTopics, fromTopicListToConjunctionTopics } = require('../shared');
 const { makeSet, makeMap, makeMultiMap } = require('../../utils/collections');
 const { trackEvent } = require('../../clients/appinsights/AppInsightsClient');
-
-function _aggregateBy(rows, aggregateKey, aggregateValue){
-  let accumulationMap = new Map();
-
-  rows.forEach(row => {
-    const key = aggregateKey(row);
-    const mapEntry = accumulationMap.has(key) ? accumulationMap.get(key) : aggregateValue(row);
-
-    const mutatedRow = Object.assign({}, mapEntry, { 
-      mentions: (mapEntry.mentions || Long.ZERO).add(row.mentioncount),
-      avgsentimentnumerator: (mapEntry.avgsentimentnumerator || Long.ZERO).add(row.avgsentimentnumerator) 
-    });
-
-    accumulationMap.set(key, mutatedRow);
-  });
-
-  return accumulationMap;
- }
 
 /**
  * @param {{site: string, timespan: string, sourceFilter: string[], mainEdge: string, originalSource: string}} args
@@ -160,6 +142,7 @@ function topSources(args, res) { // eslint-disable-line no-unused-vars
     const tilex = makeSet(tiles, tile => tile.row);
     const tiley = makeSet(tiles, tile => tile.column);
     const fetchSize = 400;
+    const responseSize = args.limit || 5;
 
     const query = `
     SELECT mentioncount, pipelinekey, externalsourceid, avgsentimentnumerator
@@ -196,15 +179,15 @@ function topSources(args, res) { // eslint-disable-line no-unused-vars
     return cassandraConnector.executeQuery(query, params, { fetchSize })
     .then(rows => {
       const filteredRows = rows.filter(row=>row.pipelinekey !== "all" || row.externalsourceid !== "all")//filter all aggregates as we're interested in named sources only
-      const accumulationMap = _aggregateBy(filteredRows, row => `${row.pipelinekey}_${row.externalsourceid}`, row => ( { 
+      const edges = 
+      aggregateBy(filteredRows, row => `${row.pipelinekey}_${row.externalsourceid}`, row => ( { 
         pipelinekey: row.pipelinekey, 
         name: row.externalsourceid, 
         mentions: Long.ZERO, 
         avgsentimentnumerator: Long.ZERO 
-      } ) );
+      } ) )
+      .slice(0, responseSize);
 
-      const edges = Array.from(accumulationMap.values()).map(source => Object.assign({}, source, { avgsentiment:  computeWeightedAvg(source.mentions, source.avgsentimentnumerator) }));
-      
       resolve({
         edges
       });
